add a function to do the initial facet reset state that is done in multiple places
also added screenshot-on-failure so we get more info about failures on CI
